+++ src/config/loader.py
@@ -1,6 +1,8 @@
 import yaml
 import os
 import project.project as proj
+import project.build as build
+import project.step as step
 
 
 class Loader:
@@ -43,7 +45,35 @@ class Loader:
 
             print("Load configs for project: '%s', host: '%s'" % (proj_configs.get("name"), proj_configs.get("host")))
 
-            self.projects.append(proj.Project(name=proj_configs.get("name"), host=proj_configs.get("host")))
+            project = proj.Project(name=proj_configs.get("name"), host=proj_configs.get("host"))
+
+            self.projects.append(project)
+
+            if 'builds' in proj_configs.keys():
+                builds_list_raw = proj_configs.get('builds')
+                assert type(builds_list_raw).__name__ == 'list', "Expected type 'list' of 'builds', but '%s' given" % type(builds_list_raw).__name__
+
+                for build_raw in builds_list_raw:
+                    assert type(build_raw).__name__ == 'dict', "Expected type 'dict' of 'builds' element, but '%s' given" % type(build_raw).__name__
+
+                    assert 'name' in build_raw.keys(), "Expected field 'name' exists in 'build[]'"
+                    assert 'description' in build_raw.keys(), "Expected field 'description' exists in 'build[]'"
+                    assert 'steps' in build_raw.keys(), "Expected field 'steps' exists in 'build[]'"
+
+                    b = build.Build(build_raw.get('name'), build_raw.get('description'))
+
+                    project.add_build(b)
+
+                    step_list_raw = build_raw.get('steps')
+                    assert type(step_list_raw).__name__ == 'list', "Expected type 'list' of 'builds.steps', but '%s' given" % type(step_list_raw).__name__
+
+                    for step_raw in step_list_raw:
+                        assert type(step_raw).__name__ == 'dict', "Expected type 'dict' of element steps, but '%s' given" % type(step_raw).__name__
+
+                        assert 'name' in step_raw.keys(), "Expected field 'name' exists"
+
+                        s = step.Step(step_raw.get('name'))
+                        b.add_step(s)
 
     def get_projects(self):
         return self.projects